Rename and extract ESS component
