+++ src/lib.rs
@@ -6,4 +6,5 @@ mod connection_provider;
 mod execution;
 mod intent_broker;
 pub use intent_broker::IntentBroker;
+pub mod ess;
 pub mod registry;
